signin-microservice/requirements.txt
flask==2.2.5
flask-jwt-extended==4.5.3
flask-session==0.4.0
flasgger==0.9.7.1  # Instead of 0.9.5
psycopg2-binary==2.9.10
redis==5.2.0
werkzeug==3.0.2
flask-cors
